requirements.txt
numpy==1.18.1
matplotlib==3.1.3
pandas==1.0.3
pytest==5.3.5
mypy==0.761
